requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-multipart
pwdlib
pyjwt